// web/src/lib/api/Group.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GroupColor } from "./GroupColor";

/**
 * An optional grouping of tasks within a project, such as a room.
 */
export type Group = { id: number, name: string, 
/**
 * `null` if the group has no color.
 */
color: GroupColor | null, };